+++ app/services/meets_service.rb
@@ -1,26 +1,42 @@
 class MeetsService
-    def initialize(meet)
-        @meet = meet
+    def initialize(meet = nil, user = nil)
+      @meet = meet
+      @user = user
     end
 
+    # Class method to mark finished meets based on date
     def self.date_check
-        Meet.mark_finished_meets
+      Meet.mark_finished_meets
     end
 
+    def self.student_meets_by_status(user, status)
+      Meet.joins(:participants)
+          .where(participants: { user_id: user.id })
+          .where(status: status)
+    end
+
+    def self.tutor_meets_by_status(user, status)
+      meets_query = Meet.joins(:availability_tutor)
+                        .where(availability_tutors: { user_id: user.id })
+      meets_query = meets_query.where(status: status) if status
+      meets_query
+    end
+
+    # Retrieves participants and topic details for a specific meet
     def get_participants_and_topic
-        query=Meet.joins("INNER JOIN availability_tutors ON availability_tutors.id = meets.availability_tutor_id")
-        .joins("INNER JOIN topics ON availability_tutors.topic_id = topics.id")
-        .joins("INNER JOIN participants ON meets.id = participants.meet_id")
-        .joins("INNER JOIN users ON participants.user_id = users.id")
-        .select("users.name AS participant_name,
-                users.email AS participant_email,
-                topics.name AS topic_name,
-                meets.description AS description,
-                meets.status AS status,
-                meets.date_time AS date_time")
-                .where(meets: { id: @meet.id })
-        query
+      query = Meet.joins("INNER JOIN availability_tutors ON availability_tutors.id = meets.availability_tutor_id")
+                  .joins("INNER JOIN topics ON availability_tutors.topic_id = topics.id")
+                  .joins("INNER JOIN participants ON meets.id = participants.meet_id")
+                  .joins("INNER JOIN users ON participants.user_id = users.id")
+                  .select("users.name AS participant_name,
+                          users.email AS participant_email,
+                          topics.name AS topic_name,
+                          meets.description AS description,
+                          meets.status AS status,
+                          meets.date_time AS date_time")
+                  .where(meets: { id: @meet.id })
+      query
     rescue ActiveRecord::RecordNotFound
-        { error: "Participants not found" }
+      { error: "Participants not found" }
     end
 end