+++ spec/requests/users_controller_spec.rb
@@ -3,6 +3,12 @@ require 'rails_helper'
 RSpec.describe "UsersControllers", type: :request do
   let!(:user) { FactoryBot.create(:user, career: nil) }
   let!(:token) { JsonWebTokenService.encode(user_id: user.id) }
+  let!(:topic) { FactoryBot.create(:topic) }
+  let!(:availability_tutor) { FactoryBot.create(:availability_tutor, user: user, topic: topic) }
+  let!(:interested) { FactoryBot.create(:interested, user: user, availability_tutor: availability_tutor) }
+  let!(:meet) { FactoryBot.create(:meet, availability_tutor: availability_tutor, status: 'confirmed') }
+  let!(:participant) { FactoryBot.create(:participant, user: user, meet: meet) }
+
   describe "GET /profile" do
     it "returns the current user with correct attributes" do
       get "/profile", headers: { 'Authorization': "Bearer #{token}" }
@@ -29,8 +35,36 @@ RSpec.describe "UsersControllers", type: :request do
     end
   end
 
-  describe "PUT /update" do
-    it "updates the current user with the correct attributes and ignore non permited attributes" do
+  describe "GET /profile/teach" do
+    it "returns the topics and meets proposed by the user as a tutor" do
+      get "/profile/teach", headers: { "Authorization" => token }
+      expect(response).to have_http_status(200)
+      body = JSON.parse(response.body)
+
+      # Check structure for teach response
+      expect(body.keys).to eq([ "meets_confirmed", "meets_pending", "meets_finished", "topics" ])
+      expect(body["meets_confirmed"].first["id"]).to eq(meet.id)
+      expect(body["topics"].first["id"]).to eq(topic.id)
+      expect(body["topics"].first["name"]).to eq(topic.name)
+    end
+  end
+
+  describe "GET /profile/learn" do
+    it "returns the topics and meets the user is interested in as a student" do
+      get "/profile/learn", headers: { "Authorization" => token }
+      expect(response).to have_http_status(200)
+      body = JSON.parse(response.body)
+
+      # Check structure for learn response
+      expect(body.keys).to eq([ "meets_confirmed", "meets_pending", "meets_finished", "topics" ])
+      expect(body["meets_confirmed"].first["id"]).to eq(meet.id)
+      expect(body["topics"].first["id"]).to eq(topic.id)
+      expect(body["topics"].first["name"]).to eq(topic.name)
+    end
+  end
+
+  describe "PUT /profile" do
+    it "updates the current user with the correct attributes and ignores non-permitted attributes" do
       put "/profile", params: { user: {
         name: "Juan Pablo II",
         description: "new description",
@@ -80,6 +114,34 @@ RSpec.describe "UsersControllers", type: :request do
         headers: { 'Authorization': "Bearer #{token}" }
       expect(response).to have_http_status(404)
       expect(JSON.parse(response.body)).to eq({ "error" => I18n.t("error.careers.not_found") })
+    end
+  end
+
+  describe "GET /users/:id/teach" do
+    it "returns the teach response for a specific user" do
+      get "/users/#{user.id}/teach", headers: { "Authorization" => token }
+      expect(response).to have_http_status(200)
+      body = JSON.parse(response.body)
+
+      # Check structure for teach response
+      expect(body.keys).to eq([ "meets_confirmed", "meets_pending", "meets_finished", "topics" ])
+      expect(body["meets_confirmed"].first["id"]).to eq(meet.id)
+      expect(body["topics"].first["id"]).to eq(topic.id)
+      expect(body["topics"].first["name"]).to eq(topic.name)
+    end
+  end
+
+  describe "GET /users/:id/learn" do
+    it "returns the learn response for a specific user" do
+      get "/users/#{user.id}/learn", headers: { "Authorization" => token }
+      expect(response).to have_http_status(200)
+      body = JSON.parse(response.body)
+
+      # Check structure for learn response
+      expect(body.keys).to eq([ "meets_confirmed", "meets_pending", "meets_finished", "topics" ])
+      expect(body["meets_confirmed"].first["id"]).to eq(meet.id)
+      expect(body["topics"].first["id"]).to eq(topic.id)
+      expect(body["topics"].first["name"]).to eq(topic.name)
     end
   end
 end